Move cleanup operation from healthcheck ctrlr to backend monitor ctrlr
