+++ internal/controller/providerconfig/healthcheck/healthcheck_controller.go
@@ -21,7 +21,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
 	v1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	"go.opentelemetry.io/otel"
@@ -40,15 +39,12 @@ import (
 	apisv1alpha1 "github.com/linode/provider-ceph/apis/v1alpha1"
 	"github.com/linode/provider-ceph/internal/consts"
 	"github.com/linode/provider-ceph/internal/otel/traces"
-	"github.com/linode/provider-ceph/internal/rgw"
 	"github.com/linode/provider-ceph/internal/utils"
 )
 
 const (
-	errHealthCheckCleanup       = "failed to perform health check cleanup"
-	errDeleteLCValidationBucket = "failed to delete lifecycle configuration validation bucket"
-	errUpdateHealthStatus       = "failed to update health status of provider config"
-	errFailedHealthCheckReq     = "failed to forward health check request"
+	errUpdateHealthStatus   = "failed to update health status of provider config"
+	errFailedHealthCheckReq = "failed to forward health check request"
 
 	healthCheckSuffix = "-health-check"
 
@@ -66,15 +62,8 @@ func (c *Controller) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	providerConfig := &apisv1alpha1.ProviderConfig{}
 	if err := c.kubeClientCached.Get(ctx, req.NamespacedName, providerConfig); err != nil {
 		if kerrors.IsNotFound(err) {
-			c.backendStore.SetBackendHealthStatus(req.Name, apisv1alpha1.HealthStatusUnknown)
-			// ProviderConfig has been deleted, perform cleanup.
-			if err := c.cleanup(ctx, req); err != nil {
-				err = errors.Wrap(err, errHealthCheckCleanup)
-				traces.SetAndRecordError(span, err)
-
-				return ctrl.Result{}, err
-			}
-
+			// ProviderConfig has been deleted so there is nothing to do and no need to requeue.
+			// The backend monitor controller will remove the backend from the backend store.
 			return ctrl.Result{}, nil
 		}
 
@@ -151,24 +140,6 @@ func (c *Controller) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	return ctrl.Result{
 		RequeueAfter: time.Duration(providerConfig.Spec.HealthCheckIntervalSeconds) * time.Second,
 	}, nil
-}
-
-// cleanup deletes the lifecycle configuration validation bucket from the backend.
-// This function is only called when a ProviderConfig has been deleted.
-func (c *Controller) cleanup(ctx context.Context, req ctrl.Request) error {
-	backendClient := c.backendStore.GetBackendS3Client(req.Name)
-	if backendClient == nil {
-		c.log.Info("Backend client not found during validation bucket cleanup - aborting cleanup", consts.KeyBackendName, req.Name)
-
-		return nil
-	}
-
-	c.log.Info("Deleting lifecycle configuration validation bucket", consts.KeyBucketName, v1alpha1.LifecycleConfigValidationBucketName, consts.KeyBackendName, req.Name)
-	if err := rgw.DeleteBucket(ctx, backendClient, aws.String(v1alpha1.LifecycleConfigValidationBucketName), true); err != nil {
-		return errors.Wrap(err, errDeleteLCValidationBucket)
-	}
-
-	return nil
 }
 
 // doHealthCheck performs a basic http request to the hostbase address.
